Fixed a bug that could cause a lockup to occur
There was one handshake that did not use a temporary variable
before checking the value in the while loop. Without the
temporary variable the XIN instruction may place the value to
be checked into a different core register than the one being
checked.

This commit adds a temporary variable in order to force the
compiler to copy the value to the variable being checked by
the while loop. All other while loops involving the XIN
instruction already used this method.

Signed-off-by: Jason Reeder <jreeder@ti.com>