docs: Excellent(!) Beagle Play guide by Bootlin folks
I wanted to archive this in case it vanishes from the web. Please note
that it is written and published by Bootlin - AFAICS it was shared under
Creative Commons - but it's better we do not directly share it.
Currently it's available at:
https://bootlin.com/doc/training/embedded-linux-beagleplay/embedded-linux-beagleplay-labs.pdf
We can refer to it if we get questions, instead of sharing this copy.
That way we also give the credit where it belongs! This copy is here
purely for back-up purposes.