remoteproc: pru: Dont enforce IRAM alignment to DRAM
It seems to be fairly easy to generate an elf which has segments which
aren't aligned by 4. At least I managed to do this by just modifying
slightly some of the pru-software-support-package/examples/am62x
examples and compiling the code with TI's toolchain.

According to the comments in remoteproc_pru file, the alginment of
4-byte boundary is an oddity of PRU IRAM. Since the unaligned segments
were going to DRAM, they should be good to go.

Relax the alignment check to concern the IRAM only.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>